move chat entity types to their own file
