+++ src/landing/model_files_item.rs
@@ -4,7 +4,7 @@ use moly_protocol::data::{File, FileID, PendingDownloadsStatus};
 use super::model_files_tags::ModelFilesTagsWidgetExt;
 use crate::{
     data::{
-        chats::chat::ChatEntityId, downloads::download::DownloadFileAction,
+        chats::chat_entity::ChatEntityId, downloads::download::DownloadFileAction,
         store::FileWithDownloadInfo,
     },
     shared::{
