+++ src/shared/actions.rs
@@ -1,7 +1,7 @@
 use makepad_widgets::{ActionDefaultRef, DefaultNone};
 use moly_protocol::data::FileID;
 
-use crate::data::chats::chat::ChatEntityId;
+use crate::data::chats::chat_entity::ChatEntityId;
 
 #[derive(Clone, DefaultNone, Debug)]
 pub enum ChatAction {
